modified file structure. moved sprites to own folder
